+++ src/bios/i386/src/dt_utils.rs
@@ -72,7 +72,7 @@ pub const ATTR_SEG16: u8 = 0b000;
 /// - 2: size bit, set if our code is 32-bit, 16-bit vice versa
 ///
 /// For granularity, CPU will multiply our limit by 4KB if this bit is set.
-pub const fn pack_dt(base: u32, limit: u32, perm: u8, s_type: u8, privilege: u8, present: u8, attrs: u8, granularity: u8) -> u64 {
+pub const fn pack_dt(base: usize, limit: usize, perm: u8, s_type: u8, privilege: u8, present: u8, attrs: u8, granularity: u8) -> u64 {
     let mut res: u64 = 0x0;
     res = mask_assign(res, limit as u64, 0, 0, 16);
     res = mask_assign(res, base as u64, 16, 0, 24);
